tests/wpt/web-platform-tests/page-visibility/unload.html

<!doctype html>
<title>visibilitychange fires on unload</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var w = window.open("resources/unload.html");
  window.onmessage = t.step_func(function(event) {
    if (event.data === "close") {
      w.close();
      return;
    }
    assert_equals(event.data, "hidden");
    t.done();
  });
});
</script>
